format total above filtered data button
