+++ src/geo/selectUtils.js
@@ -3,6 +3,19 @@ import STATEOBJ from '../constants/stateObj.js'
 import MSATONAME from '../constants/msaToName.js'
 import VARIABLES from '../constants/variables.js'
 
+function formatWithCommas(str='') {
+  str = str + ''
+  let formatted = ''
+  let comma = ','
+  for(let i = str.length; i > 0; i-=3) {
+    let start = i - 3
+    if(start < 0) start = 0
+    if(start === 0) comma = ''
+    formatted = `${comma}${str.slice(start, i)}${formatted}`
+  }
+  return formatted
+}
+
 function createStateOption(state, options){
   if(state !== 'NA') options.push({value: state, label: `${STATEOBJ[state]} - STATEWIDE`})
 }
@@ -66,5 +79,6 @@ export {
   createMSAOption,
   createGeographyOptions,
   createVariableOptions,
-  geographyStyleFn
+  geographyStyleFn,
+  formatWithCommas
 }
